[aws-lib] fold S3ClientUploader trait into it's only implementation
This trait was only implemented by S3ClientTransferManager.
