+++ aws-lib/src/main/scala/net/kemitix/s3thorp/aws/lib/S3ClientTransferManager.scala
@@ -1,21 +1,21 @@
 package net.kemitix.s3thorp.aws.lib
 
 import cats.effect.IO
+import com.amazonaws.event.{ProgressEvent, ProgressEventType, ProgressListener}
 import com.amazonaws.services.s3.model.PutObjectRequest
 import com.amazonaws.services.s3.transfer.TransferManager
 import net.kemitix.s3thorp.aws.api.S3Action.UploadS3Action
+import net.kemitix.s3thorp.aws.api.UploadEvent.{ByteTransferEvent, RequestEvent, TransferEvent}
 import net.kemitix.s3thorp.aws.api.{S3Action, UploadProgressListener}
 import net.kemitix.s3thorp.aws.lib.S3ClientTransferManagerLogging.{logMultiPartUploadFinished, logMultiPartUploadStart}
 import net.kemitix.s3thorp.domain.{Bucket, LocalFile, MD5Hash, RemoteKey}
 
-class S3ClientTransferManager(transferManager: => TransferManager)
-  extends S3ClientUploader {
+class S3ClientTransferManager(transferManager: => TransferManager) {
 
   def accepts(localFile: LocalFile)
              (implicit multiPartThreshold: Long): Boolean =
     localFile.file.length >= multiPartThreshold
 
-  override
   def upload(localFile: LocalFile,
              bucket: Bucket,
              uploadProgressListener: UploadProgressListener,
@@ -34,4 +34,18 @@ class S3ClientTransferManager(transferManager: => TransferManager)
       _ <- logMultiPartUploadFinished(localFile)
     } yield UploadS3Action(RemoteKey(result.getKey), MD5Hash(result.getETag))
   }
+
+
+  def progressListener(uploadProgressListener: UploadProgressListener): ProgressListener = {
+    new ProgressListener {
+      override def progressChanged(event: ProgressEvent): Unit = {
+        event match {
+          case e if e.getEventType.isTransferEvent => TransferEvent(e.getEventType.name)
+          case e if e.getEventType equals ProgressEventType.RESPONSE_BYTE_TRANSFER_EVENT => ByteTransferEvent(e.getEventType.name)
+          case e => RequestEvent(e.getEventType.name, e.getBytes, e.getBytesTransferred)
+        }
+      }
+    }
+  }
+
 }
